Cria controladores das classes Moradores e Utilizadores e construtor da classe Reserva
